requirements.txt
# API
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0

# LLM / Agent
langchain>=0.3.12
langchain-ollama>=0.3.0
langchain-neo4j>=0.2.0
langchain-community>=0.3.12
langchain-chroma>=0.2.0
langgraph>=0.2.60
chromadb>=0.5.0

# Database
neo4j>=5.28.0

# Production
slowapi>=0.1.9
pybreaker>=1.2.0
structlog>=24.4.0
langgraph-checkpoint-sqlite>=2.0.0
redis>=5.0.0

# Tools
mcp>=1.0.0
pyyaml>=6.0

# Testing
httpx>=0.27.0
